% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHPtools-package.R
\docType{package}
\name{AHPtools-package}
\alias{AHPtools}
\alias{AHPtools-package}
\title{AHPtools: Consistency in the Analytic Hierarchy Process}
\description{
A Swiss Army knife of utility functions for users of the Analytic Hierarchy Process (AHP) which will help you to assess the consistency of a PCM as well as to improve its consistency ratio, to compute the sensitivity of a PCM, create a logical (as distinct from a random PCM) from preferences provided for the alternatives, and a function that helps evaluate the actual consistency of a PCM based on objective, fair bench marking. The various functions in the toolkit additionally provide the flexibility to users to specify only the upper triangular comparison ratios of the PCM in order to performs its assigned task.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://CRAN.R-project.org/package=AHPtools}
}

}
\keyword{internal}
