% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHPtools.R
\name{createPCM}
\alias{createPCM}
\title{Create a Pairwise Comparison Matrix of order n for Analytic Hierarchy
Process from a vector of length n(n-1)/2 comparison ratios}
\usage{
createPCM(vec)
}
\arguments{
\item{vec}{The preference vector of length as the order of the 'PCM'}
}
\value{
A Pairwise Comparison Matrix corresponding to the upper triangular
elements
}
\description{
Create a Pairwise Comparison Matrix of order n from a vector of
length n(n-1)/2 independent upper triangular elements
}
\examples{
PCM <- createPCM(c(1,2,0.5,3,0.5,2));
PCM <- createPCM(c(1,.5,2,1/3,4,2,.25,1/3,.5,1,.2,6,2,3,1/3));
}
