% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHPtools.R
\name{createRandomPCM}
\alias{createRandomPCM}
\title{Create a Random Pairwise Comparison Matrix of given order n}
\usage{
createRandomPCM(PCMsize)
}
\arguments{
\item{PCMsize}{The order of the random 'PCM' to be generated}
}
\value{
A Pairwise Comparison Matrix corresponding to the given order
}
\description{
Create a Random Pairwise Comparison Matrix of order n for Analytic Hierarchy
Process using Saaty's 17-point Fundamental Scale for comparison ratios of the 
upper triangular elements of the order n PCM
}
\examples{
PCM <- createPCM(5);
PCM <- createPCM(11);
}
