% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitModel.R
\docType{data}
\name{logitModel}
\alias{logitModel}
\title{Logistic Regression Model}
\format{
An object of class \code{glm} containing the results of a logistic regression.
}
\source{
Generated from the dataset \code{myData}.
}
\usage{
logitModel
}
\description{
A logistic regression model object created from the \code{glm} function.
}
\examples{
data(logitModel)
summary(logitModel)
}
\keyword{datasets}
